Simplified RobotMap
I moved RobotMap into one file. It cuts down on space and compile time
to keep everything you need for RobotMap in just the .h file and fits
with the rest of your coding standards just fine. It's kinda odd to
extern a ton of variables to go define in a .cpp file - most
professional code just both will both declare and define the constants
in the .h file.